+++ packages/demiurge/vite.config.ts
@@ -9,6 +9,9 @@ export default defineConfig({
   build: {
     lib: {
       entry: {
+        "data/testing": fileURLToPath(
+          new URL("src/data/testing.ts", import.meta.url),
+        ),
         index: fileURLToPath(new URL("src/index.ts", import.meta.url)),
         "internal/testing": fileURLToPath(
           new URL("src/internal/testing.ts", import.meta.url),